Split NetworkManagerType
